portfolio-flask/requirements.txt
Flask>=2.2
Flask-Mail>=0.9.1
python-dotenv>=1.0.0
